requirements.txt
google-api-python-client
google-auth-oauthlib
groq
python-dotenv
